Small Changes
-Mainly trying to fix a screen tear bug. Otherwise just small var name changes.

+++ Objects/Battery.cs
@@ -15,7 +15,7 @@ namespace ProjectArrow.Objects
         public Battery(InputHelper inputHelper)
         {
             texture = GameData.TextureMap["Battery"];
-            position = new Vector2((int)(inputHelper.WorldMousePosition.X - texture.Width / 2), (int)(inputHelper.WorldMousePosition.Y - 12));
+            position = new Vector2((inputHelper.WorldMousePosition.X - texture.Width / 2), inputHelper.WorldMousePosition.Y - 12);
             linePosition = new Vector2(position.X + (texture.Width / 2), position.Y);
             origin = new Vector2((int)(position.X + GameData.TileSize / 2), (int)(position.Y + 12));
             depth = Helper.GetDepth(origin);
